+++ src/testing/CMakeLists.txt
@@ -16,6 +16,6 @@ include_directories(${INCLUDE_DIRECTORIES}
         ${PROJECT_SOURCE_DIR}/src
         ${PROJECT_BINARY_DIR}/src)
 
-add_library(hybridse_test_base STATIC engine_test_base.cc)
+add_library(hybridse_test_base STATIC engine_test_base.cc test_base.cc)
 target_link_libraries(hybridse_test_base hybridse_core_shared
         ${yaml_libs} ${LLVM_LIBS} ${OS_LIB} ${COMMON_LIBS} ${g_libs} ${LLVM_EXT_LIB} hybridse_flags)
